Added nested folders to new clj file creation.
New clojure files will now be created relative to the selection in
the Project window, if anything is selected.  They can also be in
nested folders, and we'll ask the compiler what the namespace
should be, given the new file's path.  File must be below the
compiler load-path, or we log a warning and don't create the file
(maybe better to create it anyway with wrong namespace value?)

Changed the #SCRIPTNAME# tag in the new file templates to
'#NAMESPACE' and '#COMPONENTNAME#', since creating new scripts no
longer relies on the DoCreateScriptAsset action, so we're not
limited to the tags it supports.