Start work on mozroots solution
Add a namespace and menu item to download Mozilla's SSL certificates
to resolve HTTPS request issues. Left out of initialization for now
because it is quite slow, and ideally we would query the system to
see if the certificates are present and only download if they are
absent. This query seems to be difficult to do, and we may need
to track it ourselves.

Waiting on confirmation of some Windows paths before merging