Improve nREPL server robustness, deal with cider reconnecting
Fixes #325.

Cider sends us packets with duplicate keys in the event of a
reconnect which BencodeNET refuses to handle. This is patched out by
checking for packets with the non-standard op "init-debugger" and ignoring
them. A more complete solution will involve switching to nREPL's own
Clojure-implemented bencode implementation, which likely supports
duplicate keys.