Add support for 'load-file' message and fix 'info' implementation
Partial support for 'load-file' is now present. This will only read the file
contents sent by cider and eval that code inside a 'do' block. This makes
'cider-eval-buffer' work.

The implementation for the 'info' message has been improved: Now it handles
symbols outside of the clojure.core namespace, and will reply with the 'no-info'
result when the var meta lookup fails.