Automatically change ns for region evals
With this changes, when an eval region command is sent, the current
ns var (*ns*) is bound to the actual ns in the file being evaled.
Detecting the ns is done by inspecting the path of the file, assuming
the folder structure maps the ns structur and the clojure's source tree
is always inside a parent folder called "Assets".

If the ns cannot be safely determined, it falls back to the previous
behaviour, and evaluates code in the last bound *ns*.